cover-letter-generator/requirements.txt
fastapi 
uvicorn 
python-docx 
python-dotenv
google-generativeai